Add some documentation for the method arb
Related to #70
From the issue description I have added some paragraphs to the README.md explaining the arb function after a primer and provided a parameter by parameter explaination of the arbs accepted params.
I found an already done example on using the arb function with Aave flash loans and provided an example of that.